pyproject.toml
[project]
name = "youtube-shorts-creator"
version = "0.1.0"
description = "AI-Powered YouTube Shorts Video Generator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.5.0",
    "streamlit>=1.46.1",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "moviepy>=2.2.1",
    "pillow>=11.2.1",
] 